remove port generation, newer k8s can handle them automcatically
